requirements.txt
Django==4.2.2
djangorestframework==3.14.0
drf-spectacular==0.26.2
psycopg2-binary==2.9.3